one fix and a couple of enhancements
Added a .gitignore file to ignore tags since they are autogenerated.

Fixed the substitute section to use a different delimiter that handles
commands that return newlines and other things better.  A lot of common
commands and subs were failing because they had slashes contained and
that would throw substitute off.  A better approach might be to use
substitute() instead.

added a global option to allow loose filetype matching so you don't have
to include an underscore in the filename to get it to match a
sub-filetype skeleton.  It's turned off by default.